Was getting an error when trying to retrieve request_token_url and access_token_url
While trying to retrieve a request token url or an access token url, I
was getting an oauth_error stating I had an invalid_signature. If you
look at the docs in
https://developer.linkedin.com/documents/authentication, for these two
seperate oauth calls, they have to be made using POST as method and not
GET. Therefore, I changed the method type to post.

After making the change, I was able to successfully retrieve the
oauth_token_secret as well as the oauth_token.